Now, the x64 linux version has problems.
It turned out that there is legitmatly wrong with the code in SDL2
and OPENGLES 2.

The only successful compilation is on the Raspberry Pi 4 Manjaro
linux with the FCMissionReader and everything that has to do with
the graphics.

I am going to try to diagnose the problem, but for now I cannot
get the graphics to work.

While I was developing, I should have checked if the graphics code
worked on my main computer more constantly.

+++ src/Data/Mission/ACTResource.cpp
@@ -66,7 +66,7 @@ Json::Value Data::Mission::ACTResource::makeJson() const {
             root["RSL"][ i ]["index"] = rsl_data[ i ].index;
     }
 
-    root["resource"]["id"] = this->matching_number;
+    root["resource"]["id"] = static_cast<unsigned int>( this->matching_number );
 
     if( tSAC.exists ) {
         root["SAC"]["unk_0"] = tSAC.unk_0;
